# openapi.yaml
openapi: 3.1.0
info:
  title: Food Bank API
  version: '1.0.0'
  description: |
    JSON API for running a food bank: referrals in, sessions scheduled, pick
    lists produced, stock adjusted.

    **Read `API.md` alongside this.** A schema describes shapes; it cannot
    describe the login/refresh cycle, what happens to a referral from an
    unrecognised address, or why a team lead receives fewer fields than an
    admin. Those are the parts most likely to be got wrong.

    ## Generating a client

    ```
    npx openapi-typescript openapi.yaml -o src/api/schema.d.ts
    ```
servers:
  - url: http://127.0.0.1:8787
    description: Local wrangler dev
  - url: https://foodbank-server-production.workers.dev
    description: Production
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/sessions/{sessionId}/confirm:
    parameters: [{ $ref: '#/components/parameters/SessionId' }]
    post:
      tags: [Pick lists]
      summary: Close the session
      description: |
        The team lead's final step. Every household must be marked either
        attended (delivered) or a no-show (not in) first: this refuses while
        anyone is still unrecorded and says which pick numbers, because a
        session closed with people unaccounted for has wrong stock figures.
        Once confirmed, the session and its pick list can no longer be changed.
        There is no override.
      responses:
        '200': { description: Confirmed }
        '409':
          description: Somebody has not been marked attended or not
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Error'
                  - type: object
                    properties:
                      error:
                        type: object
                        properties:
                          details:
                            type: object
                            properties:
                              pendingPickNumbers:
                                type: array
                                items: { type: integer }
...
  /api/v1/sessions/{sessionId}/pick-list:
    parameters: [{ $ref: '#/components/parameters/SessionId' }]
    get:
      tags: [Pick lists]
      summary: The existing session pick list
      description: >
        Read-only. POST the same path to generate or reconcile before reading.
      responses:
        '200':
          description: The pick list and its parcels
          content:
            application/json:
              schema:
                type: object
                required: [pickList, parcels]
                properties:
                  pickList: { $ref: '#/components/schemas/PickList' }
                  parcels:
                    type: array
                    items: { $ref: '#/components/schemas/Parcel' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      tags: [Pick lists]
      summary: Generate or reconcile the pick list
      description: |
        Idempotent — calling again creates parcels for the session's households
        holding a place (`pending_review`, `active` and `reviewed`) which do not
        already have one, and reports that number in `parcelsCreated`. Existing
        parcels are never changed: their household
        snapshot and any manual line changes stay intact. Once confirmed, the
        list is locked and creates nothing.

        A non-zero `parcelsCreated` after the list was printed means the
        client must offer to print again so the new households have sheets.
        Contents are **copied** from the model parcels, so later rule edits
        cannot change any existing parcel.

        Generation is all-or-nothing. If any household due a parcel has no
        configured model parcel, no new parcels are created and the administrator must
        complete the household grid first.
      responses:
        '200':
          description: Generated, reconciled, or already complete
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PickList'
                  - type: object
                    properties:
                      parcelsCreated: { type: integer }
                      linesCreated: { type: integer }
        '422': { description: A model parcel is not configured for every booked household size }
  /api/v1/pick-lists/{id}:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    get:
      tags: [Pick lists]
      summary: A pick list with its parcels
      responses:
        '200':
          description: The pick list
          content:
            application/json:
              schema:
                type: object
                required: [pickList, parcels]
                properties:
                  pickList: { $ref: '#/components/schemas/PickList' }
                  parcels:
                    type: array
                    items: { $ref: '#/components/schemas/Parcel' }
  /api/v1/pick-lists/{id}/print:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    get:
      tags: [Pick lists]
      summary: The printable payload
      description: |
        One sheet per parcel, lines ordered by shelf. JSON — the frontend owns
        layout.

        **Never contains the reason for referral**, and no name or address
        unless it is a delivery. A sheet gets carried round a hall and left on
        tables. Dietary notes *are* included: the picker is the only person who
        can act on them.
      responses:
        '200':
          description: Sheets
          content:
            application/json:
              schema:
                type: object
                required: [pickList, parcels]
                properties:
                  pickList: { $ref: '#/components/schemas/PickList' }
                  parcels:
                    type: array
                    items: { $ref: '#/components/schemas/PrintParcel' }
    post:
      tags: [Pick lists]
      summary: Mark as printed
      description: >
        Only the first print is stamped. **Lines can still be edited after
        printing** — the list locks on confirm, not on print.
      responses:
        '200':
          description: Marked
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PickList' }
        '409': { description: Already confirmed }
  /api/v1/pick-lists/{id}/confirm:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    post:
      tags: [Pick lists]
      summary: Lock the pick list
      description: |
        Picking is finished and the list can no longer be edited.

        **This does not move stock.** Stock moves when attendance is recorded,
        because until someone turns up nothing has been given away. Idempotent.
      responses:
        '200':
          description: Confirmed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PickList' }
  /api/v1/pick-lists/{id}/divergence:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    get:
      tags: [Pick lists]
      summary: What still differs from the pick list
      description: >
        Household-size changes and cancelled referrals are reported, never
        applied automatically. While the list is editable, POSTing its session
        pick-list reconciles referrals that have no parcel yet — `pending_review`,
        `active` and `reviewed` alike; a confirmed list remains locked and can still report
        missing referrals.
      responses:
        '200':
          description: Differences
          content:
            application/json:
              schema:
                type: object
                properties:
                  missingParcels:
                    type: array
                    items: { type: string, format: uuid }
                    description: Households holding a place without a parcel. An editable list
                      reconciles these when its session pick-list is POSTed; a confirmed
                      list remains locked.
                  changedHouseholds:
                    type: array
                    items:
                      type: object
                      properties:
                        parcelId: { type: string, format: uuid }
                        was: { type: object }
                        now: { type: object }
                  cancelledReferrals:
                    type: array
                    items: { type: string, format: uuid }
  /api/v1/parcels/{id}:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    patch:
      tags: [Pick lists]
      summary: Set a note on a parcel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [notes]
              properties:
                notes: { type: [string, 'null'], maxLength: 500 }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                type: object
                required: [id, isActive]
                properties:
                  id: { type: string, format: uuid }
                  isActive: { type: boolean }
        '409': { description: The pick list has been confirmed }
  /api/v1/parcels/{id}/lines:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    put:
      tags: [Pick lists]
      summary: Add an item or change its quantity
      description: |
        Marked as a manual change so it is distinguishable from what the model
        parcel decided. **A quantity of 0 removes the line**, which is how a
        picker records "we had none".
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [stockItemId, quantity]
              properties:
                stockItemId: { type: string, format: uuid }
                quantity: { type: integer, minimum: 0, maximum: 1000 }
      responses:
        '204': { description: Applied }
        '409': { description: The pick list has been confirmed }
  /api/v1/parcels/{id}/review:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    post:
      tags: [Pick lists]
      summary: Mark a parcel's pick list reviewed
      responses:
        '200':
          description: Reviewed
          content:
            application/json:
              schema:
                type: object
                required: [id, reviewedAt]
                properties:
                  id: { type: string, format: uuid }
                  reviewedAt: { type: string, format: date-time }
        '409': { description: The pick list has been confirmed }
  /api/v1/parcels/{id}/lines/{stockItemId}:
    parameters:
      - $ref: '#/components/parameters/Id'
      - { in: path, name: stockItemId, required: true, schema: { type: string, format: uuid } }
    delete:
      tags: [Pick lists]
      summary: Remove an item from a parcel
      responses:
        '204': { description: Removed }
        '409': { description: The pick list has been confirmed }
  /api/v1/parcels/{id}/attendance:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    post:
      tags: [Pick lists]
      summary: Did this household turn up?
      description: |
        **Attended** issues the parcel and decrements stock. **No-show** moves
        nothing — the parcel is unpacked and nothing was ever given away. For a
        delivery these are the same two outcomes under the driver's names:
        delivered and not in.

        **The parcel's pick list must have been reviewed first.** Call
        `POST /parcels/{id}/review` before this, or it is refused with a `409`
        — an outcome recorded against a list nobody checked moves stock that may
        not match what went in the bag. `Parcel.reviewedAt` tells you whether it
        has been.

        Safe to submit repeatedly: `alreadyRecorded: true` means the outcome was
        already in place and stock did not move again.

        **An outcome can be taken back while the session is open.** Marking a
        household a no-show after marking them attended deletes that parcel's
        stock movements and puts the goods back — nothing left the building, so
        nothing should have come off the shelves. Marking them attended again
        takes it again. Flip as often as needed; the level is always the sum of
        what is actually there.

        This is the only way to fix a mis-tap: the hand correction that used to
        do it went with the stock simplification.

        **Confirming the session ends it.** After `POST /sessions/{id}/confirm`
        an outcome can no longer be changed and this returns `409`, because a
        session that has been signed off must not have its figures move
        underneath it.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [attendance]
              properties:
                attendance: { type: string, enum: [attended, no_show] }
      responses:
        '200':
          description: Recorded
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string, format: uuid }
                  attendance: { type: string }
                  stockMoved: { type: boolean }
                  alreadyRecorded: { type: boolean }
        '409':
          description: >-
            The pick list has not been reviewed yet, the session has been
            confirmed, or somebody else recorded this household at the same
            moment.

  # ---------------------------------------------------------- admin set-up --
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        `Authorization: Bearer <accessToken>`. Expires after 15 minutes, or
        when the eight-hour sign-in ends if that comes first; see the refresh
        flow in API.md.
  parameters:
    Id:
      in: path
      name: id
      required: true
      schema: { type: string, format: uuid }
    SessionId:
      in: path
      name: sessionId
      required: true
      schema: { type: string, format: uuid }
  responses:
...
    NotFound:
      description: No such thing.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      required: [error]
      description: >
        Every failure has this shape. `requestId` also appears in the
        `x-request-id` header — quote it when reporting a problem.
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code:
              type: string
              enum:
                [
                  BAD_REQUEST,
                  VALIDATION_FAILED,
                  UNAUTHORIZED,
                  FORBIDDEN,
                  NOT_FOUND,
                  CONFLICT,
                  UNPROCESSABLE,
                  INTERNAL_ERROR,
                ]
            message:
              type: string
              description: Safe to show a user. Never contains personal data.
            details:
              type: object
              description: Structured context, such as which fields failed.
            requestId: { type: string }
...
    PickList:
      type: object
      required: [id, sessionId, status, generatedAt, firstPrintedAt, confirmedAt]
      properties:
        id: { type: string, format: uuid }
        sessionId: { type: string, format: uuid }
        status: { type: string, enum: [draft, printed, confirmed] }
        generatedAt: { type: string, format: date-time }
        firstPrintedAt: { type: [string, 'null'], format: date-time }
        confirmedAt: { type: [string, 'null'], format: date-time }
    ParcelLine:
      type: object
      required: [stockItemId, name, shelfNumber, quantity, source]
      properties:
        stockItemId: { type: string, format: uuid }
        name: { type: string }
        shelfNumber: { type: string }
        quantity: { type: integer }
        source:
          type: string
          enum: [model, manual]
          description: '`manual` means a human changed it after generation.'
    Parcel:
      type: object
      required:
        [
          id,
          referralId,
          pickNumber,
          refereeFirstName,
          refereeSurname,
          isDelivery,
          adults,
          children,
          householdSize,
          reviewedAt,
          attendance,
          notes,
          answers,
          lines,
        ]
      properties:
        id: { type: string, format: uuid }
        referralId: { type: string, format: uuid }
        pickNumber:
          type: integer
          description: The number on the printed sheet, 1..N within the session.
        refereeFirstName:
          type: [string, 'null']
          description: The household name for the session workspace; never a referral reason.
        refereeSurname: { type: [string, 'null'] }
        isDelivery:
          type: boolean
          description: Chooses delivery attendance labels. Address and phone stay out of this response.
        adults: { type: integer }
        children: { type: integer }
        householdSize: { type: integer }
        reviewedAt: { type: [string, 'null'], format: date-time }
        attendance: { type: string, enum: [pending, attended, no_show, cancelled] }
        notes: { type: [string, 'null'] }
        answers:
          type: object
          additionalProperties: true
          description: >
            The referral's answers, **whole and unfiltered**, for the
            preferences half of the pick-list maintenance screen.

            Which of them are preferences is yours to know: you own the form
            definition and the `preference` flag on each question, so filter
            this map yourself. The server holds no definition and will not
            guess — the four hard-coded dietary keys this replaced were exactly
            that guess, and none of them is a key in the real form.

            Empty once the referral's personal data has been purged.
        lines:
          type: array
          items: { $ref: '#/components/schemas/ParcelLine' }
    PrintParcel:
      type: object
      description: >
        Deliberately narrow. **No reason for referral, ever**, and no answers —
        by print time the decision they informed is in `lines`.

        The pick number and the household's name are on every sheet: the person
        carrying the bag has to hand it to somebody. The address, postcode and
        phone number appear **only** for a delivery.
      required:
        [
          pickNumber,
          householdSize,
          adults,
          children,
          refereeFirstName,
          refereeSurname,
          isDelivery,
          deliveryAddress,
          deliveryPostcode,
          deliveryPhone,
          notes,
          lines,
        ]
      properties:
        pickNumber: { type: integer }
        householdSize: { type: integer }
        adults: { type: integer }
        children: { type: integer }
        refereeFirstName: { type: [string, 'null'] }
        refereeSurname: { type: [string, 'null'] }
        isDelivery:
          type: boolean
          description: Print the word `DELIVERY` when this is true.
        deliveryAddress:
          type: [string, 'null']
          description: >-
            Where to drive to. This is the referee's own address — a delivery
            never goes anywhere else — and is null for a collection.
        deliveryPostcode: { type: [string, 'null'] }
        deliveryPhone:
          type: [string, 'null']
          description: For the driver who cannot find the door. Null for a collection.
        notes: { type: [string, 'null'] }
        lines:
          type: array
          items: { $ref: '#/components/schemas/ParcelLine' }

    # ---- Admin setup: referrers and reasons ----
    # All mapped by exhaustive object literals, so every field is required.
...